requirements.txt
fastapi
tortoise-orm[asyncpg]
uvicorn
aerich
pydantic[email]
pydantic[dotenv]
pydantic-settings
passlib[bcrypt]
pyjwt[crypto]

git+https://github.com/HexChap/MSCore#egg=ms_core